Split up http.go.
