+++ apiserver/common/http/endpoint_test.go
@@ -1,0 +1,37 @@
+// Copyright 2016 Canonical Ltd.
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+package http_test
+
+import (
+	//jc "github.com/juju/testing/checkers"
+	gc "gopkg.in/check.v1"
+
+	//"github.com/juju/juju/apiserver/common/http"
+)
+
+type EndpointSpecSuite struct {
+	BaseSuite
+}
+
+var _ = gc.Suite(&EndpointSpecSuite{})
+
+func (s *EndpointSpecSuite) TestNewEndpointSpec(c *gc.C) {
+	// TODO(ericsnow) This needs to be implemented ASAP.
+}
+
+func (s *EndpointSpecSuite) TestAdd(c *gc.C) {
+	// TODO(ericsnow) This needs to be implemented ASAP.
+}
+
+func (s *EndpointSpecSuite) TestPattern(c *gc.C) {
+	// TODO(ericsnow) This needs to be implemented ASAP.
+}
+
+func (s *EndpointSpecSuite) TestMethods(c *gc.C) {
+	// TODO(ericsnow) This needs to be implemented ASAP.
+}
+
+func (s *EndpointSpecSuite) TestResolve(c *gc.C) {
+	// TODO(ericsnow) This needs to be implemented ASAP.
+}
