+++ apiserver/common/http/endpoints_test.go
@@ -1,0 +1,37 @@
+// Copyright 2016 Canonical Ltd.
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+package http_test
+
+import (
+	//jc "github.com/juju/testing/checkers"
+	gc "gopkg.in/check.v1"
+
+	//"github.com/juju/juju/apiserver/common/http"
+)
+
+type EndpointsSuite struct {
+	BaseSuite
+}
+
+var _ = gc.Suite(&EndpointsSuite{})
+
+func (s *EndpointsSuite) TestNewEndpoints(c *gc.C) {
+	// TODO(ericsnow) This needs to be implemented ASAP.
+}
+
+func (s *EndpointsSuite) TestAdd(c *gc.C) {
+	// TODO(ericsnow) This needs to be implemented ASAP.
+}
+
+func (s *EndpointsSuite) TestSpecs(c *gc.C) {
+	// TODO(ericsnow) This needs to be implemented ASAP.
+}
+
+func (s *EndpointsSuite) TestResolve(c *gc.C) {
+	// TODO(ericsnow) This needs to be implemented ASAP.
+}
+
+func (s *EndpointsSuite) TestResolveForMethods(c *gc.C) {
+	// TODO(ericsnow) This needs to be implemented ASAP.
+}
